Removed the given Starter class, and established basic Model/View/Controller pattern
The view, being the main window, is the new TKGasketDriver class. It is
mainly adopted from the Gasket demo, with a few additions such as time
keeping.

TKScene should be the base class for all scene controllers. It defines
two API for drawing a scene and update a scene's states. The states can
be considered as part of the Model domain